databinding

+++ app/src/main/java/com/example/weatherapp/model/condition/Imperial.kt
@@ -14,7 +14,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Imperial (
 
-	@SerializedName("Value") val value : Int,
+	@SerializedName("Value") val value : Double,
 	@SerializedName("Unit") val unit : String,
 	@SerializedName("UnitType") val unitType : Int
 )
